add `react/internal` entry point, reorganize imports
