--- src/react/cache/index.ts
@@ -1,2 +1,0 @@
-export type { SuspenseCacheOptions } from "./SuspenseCache.js";
-export { getSuspenseCache } from "./getSuspenseCache.js";
