+++ src/react/hooks/useBackgroundQuery.ts
@@ -11,16 +11,15 @@ import {
   unwrapQueryRef,
   updateWrappedQueryRef,
   wrapQueryRef,
-} from "../cache/QueryReference.js";
-import type { QueryReference } from "../cache/QueryReference.js";
+  getSuspenseCache,
+} from "../internal/index.js";
+import type { QueryReference, CacheKey } from "../internal/index.js";
 import type { BackgroundQueryHookOptions, NoInfer } from "../types/types.js";
 import { __use } from "./internal/index.js";
-import { getSuspenseCache } from "../cache/index.js";
 import { useWatchQueryOptions } from "./useSuspenseQuery.js";
 import type { FetchMoreFunction, RefetchFunction } from "./useSuspenseQuery.js";
 import { canonicalStringify } from "../../cache/index.js";
 import type { DeepPartial } from "../../utilities/index.js";
-import type { CacheKey } from "../cache/types.js";
 import type { SkipToken } from "./constants.js";
 
 export type UseBackgroundQueryResult<
