+++ src/react/hooks/useLoadableQuery.ts
@@ -11,11 +11,11 @@ import {
   unwrapQueryRef,
   updateWrappedQueryRef,
   wrapQueryRef,
-} from "../cache/QueryReference.js";
-import type { QueryReference } from "../cache/QueryReference.js";
+  getSuspenseCache,
+} from "../internal/index.js";
+import type { QueryReference, CacheKey } from "../internal/index.js";
 import type { LoadableQueryHookOptions } from "../types/types.js";
 import { __use, useRenderGuard } from "./internal/index.js";
-import { getSuspenseCache } from "../cache/index.js";
 import { useWatchQueryOptions } from "./useSuspenseQuery.js";
 import type { FetchMoreFunction, RefetchFunction } from "./useSuspenseQuery.js";
 import { canonicalStringify } from "../../cache/index.js";
@@ -23,7 +23,6 @@ import type {
   DeepPartial,
   OnlyRequiredProperties,
 } from "../../utilities/index.js";
-import type { CacheKey } from "../cache/types.js";
 import { invariant } from "../../utilities/globals/index.js";
 
 export type LoadQueryFunction<TVariables extends OperationVariables> = (
